move to traits
