+++ components/script/dom/webgpu/gputextureview.rs
@@ -6,7 +6,7 @@ use dom_struct::dom_struct;
 use js::context::{JSContext, NoGC};
 use script_bindings::cell::DomRefCell;
 use script_bindings::reflector::{Reflector, reflect_dom_object_with_cx};
-use script_webgpu::promise::GPUTextureViewTrait;
+use script_webgpu::traits::GPUTextureViewTrait;
 use webgpu_traits::{WebGPU, WebGPURequest, WebGPUTextureView};
 
 use crate::dom::bindings::codegen::Bindings::WebGPUBinding::GPUTextureViewMethods;
